Refactoring
- Rename QualifiedEnumArgumentFactory to EnumArgumentFactory
- Made EnumStrategy top level class
- Moved EnumByName, EnumByOrdinal, Enums, and EnumStrategy to
  org.jdbi.v3.core.enums
- Removed Enums.defaultByName() and .defaultByOrdinal() in favor of
  .setDefaultStrategy(EnumStrategy)
- Moved Enums.findStrategy() to separate config class in an internal
  package.
- Rename QualifiedEnumMapperFactory to EnumMapperFactory, move it to
  org.jdbi.v3.core.mapper, and make it package private
- Un-deprecate EnumMapper, and move byName() and byOrdinal()
  static factory methods back from EnumMapperFactory to EnumMapper.
- Add API alternative messages to EnumByOrdinalMapperFactory
  and EnumByNameMapperFactory
- Use an expiring cache for EnumMapper.byName()
