+++ core/src/main/java/org/jdbi/v3/core/enums/EnumByName.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.v3.core;
+package org.jdbi.v3.core.enums;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
