+++ core/src/main/java/org/jdbi/v3/core/mapper/ColumnMappers.java
@@ -24,7 +24,6 @@ import org.jdbi.v3.core.config.ConfigRegistry;
 import org.jdbi.v3.core.config.JdbiConfig;
 import org.jdbi.v3.core.generic.GenericType;
 import org.jdbi.v3.core.internal.JdbiOptionals;
-import org.jdbi.v3.core.internal.QualifiedEnumMapperFactory;
 import org.jdbi.v3.core.qualifier.QualifiedType;
 import org.jdbi.v3.meta.Beta;
 
@@ -46,7 +45,7 @@ public class ColumnMappers implements JdbiConfig<ColumnMappers> {
         register(new BoxedMapperFactory());
         register(new PrimitiveMapperFactory());
         register(new OptionalMapperFactory());
-        register(new QualifiedEnumMapperFactory());
+        register(new EnumMapperFactory());
         register(new NVarcharMapper());
     }
