+++ core/src/test/java/org/jdbi/v3/core/EnumsConfigTest.java
@@ -15,6 +15,11 @@ package org.jdbi.v3.core;
 
 import java.lang.annotation.RetentionPolicy;
 
+import org.jdbi.v3.core.enums.EnumByName;
+import org.jdbi.v3.core.enums.EnumByOrdinal;
+import org.jdbi.v3.core.enums.EnumStrategy;
+import org.jdbi.v3.core.enums.Enums;
+import org.jdbi.v3.core.internal.EnumStrategies;
 import org.jdbi.v3.core.qualifier.QualifiedType;
 import org.jdbi.v3.core.result.UnableToProduceResultException;
 import org.jdbi.v3.core.rule.SqliteDatabaseRule;
@@ -31,7 +36,7 @@ public class EnumsConfigTest {
     @Test
     public void byNameIsDefault() {
         assertThat(db.getJdbi().getConfig(Enums.class).getDefaultStrategy())
-            .isEqualTo(Enums.EnumStrategy.BY_NAME);
+            .isEqualTo(EnumStrategy.BY_NAME);
     }
 
     @Test
@@ -68,7 +73,7 @@ public class EnumsConfigTest {
     @Test
     public void ordinalsAreBoundCorrectly() {
         db.getJdbi().useHandle(h -> {
-            h.getConfig(Enums.class).defaultByOrdinal();
+            h.getConfig(Enums.class).setEnumStrategy(EnumStrategy.BY_ORDINAL);
 
             h.createUpdate("create table enums(id int, ordinal int)").execute();
 
@@ -88,7 +93,7 @@ public class EnumsConfigTest {
     @Test
     public void ordinalsAreMappedCorrectly() {
         db.getJdbi().useHandle(h -> {
-            h.getConfig(Enums.class).defaultByOrdinal();
+            h.getConfig(Enums.class).setEnumStrategy(EnumStrategy.BY_ORDINAL);
 
             Foobar name = h.createQuery("select :ordinal")
                 .bind("ordinal", Foobar.FOO.ordinal())
@@ -112,7 +117,7 @@ public class EnumsConfigTest {
     @Test
     public void badOrdinalThrows() {
         db.getJdbi().useHandle(h -> {
-            h.getConfig(Enums.class).defaultByOrdinal();
+            h.getConfig(Enums.class).setEnumStrategy(EnumStrategy.BY_ORDINAL);
 
             assertThatThrownBy(h.createQuery("select 2").mapTo(Foobar.class)::findOnly)
                 .isInstanceOf(UnableToProduceResultException.class)
@@ -150,13 +155,13 @@ public class EnumsConfigTest {
     public void testConflictingQualifiers() {
         QualifiedType<RetentionPolicy> type = QualifiedType.of(RetentionPolicy.class).with(EnumByName.class, EnumByOrdinal.class);
 
-        assertThatThrownBy(() -> db.getJdbi().getConfig(Enums.class).findStrategy(type, RetentionPolicy.class))
+        assertThatThrownBy(() -> db.getJdbi().getConfig(EnumStrategies.class).findStrategy(type))
             .isInstanceOf(IllegalArgumentException.class);
     }
 
     @Test
     public void testConflictingSourceAnnotations() {
-        assertThatThrownBy(() -> db.getJdbi().getConfig(Enums.class).findStrategy(QualifiedType.of(BiPolar.class), BiPolar.class))
+        assertThatThrownBy(() -> db.getJdbi().getConfig(EnumStrategies.class).findStrategy(QualifiedType.of(BiPolar.class)))
             .isInstanceOf(IllegalArgumentException.class);
     }
